M33603: Adding "to limit"
Hello @iainfoulds,
Translator has reported possible source content issue.
Description: This string is grammatically correct, but does not completely make sense to me. A verb (and a preposition "to") might be missing after "and", like "and to set/define/change the permissions they then have...". I guess "set" or "define" is the missing one, but translated the source string like "To limit who can get that Kubernetes configuration ({1}kubeconfig{2}) information and to limit the permissions they then have..." as per the source.
Please review and merge the suggested proposed file change to avoid this error. If you make related fix in another PR, then share your PR number, so we can confirm and close this PR.
Many thanks in advance.