Add summary table of authorization options
This is to clarify confusion which I noticed in a recent customer support issue.  They (and I) couldn't piece together the full picture, since different parts of it were documented in different places. This new table brings together the full picture of all options, and links out to the various other pages.