Fixing the RG where ANF account is created.
Following the official Netapp documentation for Azure here:
https://docs.netapp.com/us-en/astra-control-service/get-started/set-up-microsoft-azure-with-anf.html#create-a-netapp-account-2
and here https://docs.microsoft.com/en-us/azure/azure-netapp-files/azure-netapp-files-create-netapp-account
there's no REQUIREMENT for the account to be created in
the MC resource of AKS.

The only requirement in the docs is related to the subnet delegation.

Having the ANF created in the MC gives the impression that the
service is tied "artificially" to one AKS cluster, whereas the ANF
service can be used accross multiple AKS clusters.