Update csi-secrets-store-nginx-tls.md
I found 5 mistakes on the yamls file format, and the yaml files have been fixed.

on file aks-helloworld-one.yaml and aks-helloworld-two.yaml on the service section it will not create because the format was wrong 

  ports:
  - port: 80
the - it was before the ports:

and on the ingress file there was two mistakes :

  tls:
 - hosts:
should be. 

  tls:
  - hosts:

and also :

  rules:
 - host: demo.azure.com
should be 

  rules:
  - host: demo.azure.com

now the mistakes fixed to make sure once the customer trying to run those yaml files without an issues.