Fix missing parts and refine workload-identity-deploy-cluster.md
1. Modify "-otsv" to "-o tsv"

I understand these commands can be executed, but I am not sure if it supposed to be running with "-otsv", it should not be supposed to be like this. This is too weird.  

2. Refine the description of OIDC_ISSUER

Give more detail to prevent misunderstanding, like believing "/00000000-0000-0000-0000-000000000000/11111111-1111-1111-1111-111111111111/" is represting oidc key.  

3. Refine the part of "Deploy your application"
Even don't want to give exact image name, it is too awful to give an incorrect/substandard yaml.  

4. Refine the part "Optional - Grant permissions to access Azure Key Vault"

a. We have set environment with same name before. So delete/refine the repeating part. Or user may overwrite the environment parameter by accident. 
b. It looks like this part is incomplete. There is no way there is only one step. 
c. Am I looking at an article presenting in the disorder way? Why the steps to verify tokens is in the last section?