Improve steps order and notes in deploy-confidential-containers-default-policy.md
1. Change some of the content  
For example: create kafka namespace should be done when create serviceaccount of workload identity. No matter it is on existing AKS or new AKS, there is no way it can be done in the middle of setting demo.  

2. Add some necessary notes  
One key part is the environment variable `MANAGED_IDENTITY`, this is required by bash script but not listed.  
  
3. Improve the order  
The yaml part can be done at the last. We need to replace content in the yaml files anyway. So why ask user to create it first then modify it? It should be better if we have done everything and change it in one shot.