Spread masters in at least 2 zones.
https://redis.io/docs/latest/operate/oss_and_stack/management/scaling/#creating-and-using-a-redis-cluster
Redis Cluster provides a way to run a Redis installation where data is automatically sharded across multiple Redis nodes. Redis Cluster also provides some degree of availability during partitions—in practical terms, the ability to continue operations when some nodes fail or are unable to communicate. However, the cluster will become unavailable in the event of larger failures (for example, when the majority of masters are unavailable).