Update deploy-valkey-cluster.md to improve readability and navigation
Rather than stating "Deploy" I added H2 Headers for each new step since code blocks were so long. Since this is the configure and deploy doc, I think this change should fit even if we don't explicitly say "deploy" in the doc's. This creates links in the page so that users can easily navigate to the new section. It also should improve readability because they will clearly see which section they are in.

Also added a summary in the intro.